steamcrawler/requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
bcrypt==3.1.7
beautifulsoup4==4.7.1
botocore==1.12.176
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
docutils==0.14
html2text==2018.1.9
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jmespath==0.9.4
lxml==4.3.4
numpy==1.16.4
packaging==19.0
pandas==0.24.2
paramiko==2.6.0
parsel==1.5.1
pkg-resources==0.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
queuelib==1.5.0
requests==2.22.0
Scrapy==1.6.0
selectolax==0.1.13
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
smart-getenv==1.1.0
soupsieve==1.9.2
stem==1.7.1
Twisted==19.2.1
urllib3==1.25.3
w3lib==1.20.0
warc==0.2.1
zope.interface==4.6.0
